% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFTC_leaf_area_EBImage.R
\name{leaf_area_calculation}
\alias{leaf_area_calculation}
\title{Calculate leaf area from a folder of jpeg images}
\usage{
leaf_area_calculation(folder_path, output_folder, method = "Otsu")
}
\arguments{
\item{folder_path}{Path to the folder containing JPEG images}

\item{output_folder}{Path to the folder where output files will be saved}

\item{method}{Method for thresholding (default is "Otsu")}
}
\value{
A data frame containing the results of leaf area calculation
}
\description{
Calculate leaf area from a folder of jpeg images
}
